% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggwrap.R
\name{ggwrap}
\alias{ggwrap}
\title{wrap ggbeast plots}
\usage{
ggwrap(...)
}
\arguments{
\item{...}{one or more ggbeast plots separeted by collom.}
}
\description{
This function grabs the data from independent plots and create an unique dataset for using with regularized aesthetics.
}
\seealso{
\code{\link[=harvest_ebsp]{harvest_ebsp()}}, \code{\link[=scale_time]{scale_time()}}, \code{\link[=scale_Ne]{scale_Ne()}}, \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}
}
\author{
Wilson Frantine-Silva
}
