% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harves_ebsp.R
\name{hasvest_ebsp}
\alias{hasvest_ebsp}
\alias{harvest_ebsp}
\title{process EBSP log file to data_frame}
\usage{
harvest_ebsp(path = NULL, burnin = 0.1)
}
\arguments{
\item{path}{the file path to EBSP log or a data.frame from read table with the expected structure. To see that, take a look at data("ebsp") output.}

\item{burnin}{the amount of runs to discard from 0 (keep all) to 1 (discard all and result in a error). The Default is 0.1 (i.e. 10\%).

This function also add an aesthetic layer to the data to help users to manipulate characteristics for each run.
Thus, each ggplot2 object data will have an extra column at $data slot with the name of the file that generated the data.

This function is based on the work of Joseph Helled}
}
\description{
extract the main parameters from EBSP.log file and output a data.frame
}
\examples{
\dontrun{
path <- system.file("extdata/ebsp.txt", package="ggbeast")
harvest_ebsp(path) |> head()
}
}
\references{
BEAST2 Tutorials Extended Bayesian Skyline Plot.
}
