% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_Ne}
\alias{scale_Ne}
\alias{scale_Ne_log}
\alias{scale_Ne_linear}
\title{Scale Effective Population Size (Ne) axis}
\usage{
scale_Ne_log()

scale_Ne_linear()
}
\description{
This family of functions are just alias to scale_y_log10 and scale_y_continuous() with more intuitive names.
}
