% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggebsp.R
\name{ggebsp}
\alias{ggebsp}
\title{Plot EBSP with ggplot2}
\usage{
ggebsp(file = NULL, burnin = 0.1)
}
\arguments{
\item{file}{the path to EBSP.log run file}

\item{burnin}{the burnin quantity parameter. By defaut it is 0.1 (10\%), but users may adjust as tracer output}
}
\description{
this function outputs an ggplot2 object for a Extended Bayesian Skyline plot output as described in details in the EBSP BEAST2 tutorial by Joseph Helled
}
\author{
Wilson Frantine-Silva
}
